app/requirements.txt
fastapi
uvicorn
streamlit
langchain
mysql-connector-python
openai
langchain
langchain-openai
langchain_community
langchain-community
langsmith
python-dotenv
pandas
passlib
pydantic
pyjwt


